backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
psycopg[binary]>=3.1
django-cors-headers